% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{ammonia}
\alias{ammonia}
\title{Ammonia size data}
\format{A data frame with 867 rows and 7 variables:
\describe{
  \item{species}{species of genus Ammonia from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{h}{test height in μm (82.05--390.65)}
  \item{d_one}{test minor diameter in  μm (20--220)}
  \item{d_two}{test major diameter in  μm (75.1--364.6)}
  \item{area}{test surface area in μm2 (4438--107903)}
  \item{pco}{percent cell occupancy of the test (11--100)}
}}
\usage{
ammonia
}
\description{
A dataset containing the size and other attributes of species from genus Ammonia.
 The variables are as follows:
}
\keyword{datasets}
