% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{amphistegina}
\alias{amphistegina}
\title{\emph{Amphistegina} size data}
\format{
A data frame with 167 rows and 5 variables:
\describe{
  \item{ind}{number of individuals}
  \item{h}{test height in μm (60.0--579.0)}
  \item{d_one}{test diameter in  μm (111.3--1193.1)}
  \item{area}{test surface area in μm2 (9722--3027282)}
  \item{pco}{percent cell occupancy of the test (3.408--100)}
}
}
\usage{
amphistegina
}
\description{
A dataset containing the size and other attributes of genus \emph{Amphistegina}.
 The variables are as follows:
}
\keyword{datasets}
