% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{asterotrochammina}
\alias{asterotrochammina}
\title{Asterotrochammina size data}
\format{A data frame with 335 rows and 7 variables:
\describe{
  \item{species}{species of genus Asterotrochammina from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{h}{test height in μm (12--72)}
  \item{d_one}{test minor diameter in  μm (77.86--265.50)}
  \item{radius}{test radius in  μm (38.93--132.75)}
  \item{area}{test surface area in μm2 (4761--57577)}
  \item{pop}{test protoplasm occupancy percentage (15--100)}
}}
\usage{
asterotrochammina
}
\description{
A dataset containing the size and other attributes of species from genus Asterotrochammina.
 The variables are as follows:
}
\keyword{datasets}
