% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{data_pco}
\alias{data_pco}
\title{Foraminiferal genera data for \code{forImage} examples}
\format{
A `data.frame` with 72 rows and 3 variables:
\describe{
  \item{genera}{foraminifera genera}
  \item{mean}{mean protoplasm occupancy percentage \code{(pco)}}
  \item{sd}{standard deviation of mean \code{(pco)}}
  \item{n}{number of specimens}
  \item{model}{fitted geometric model in Freitas and others (2019)}
}
}
\description{
This `data.frame` contains protoplasm occupancy mean and standard deviation data of 72 foraminifera genera.
These genera data are originally from Freitas and others (2019)
}
\details{
Foraminifera genera with \code{(pco)} available:
\tabular{rlll}{
 \tab"ammonia"\tab"amphistegina"\tab"angulogerina"\cr
 \tab"archaias"\tab"asterotrochammina"\tab"bolivina"\cr
 \tab"buliminella"\tab"cancris"\tab"caronia"\cr
 \tab"cassidulina"\tab"cibicides"\tab"cibicidoides"\cr
 \tab"cornuspira"\tab"cribroelphidium"\tab"cymbaloporetta"\cr
 \tab"deuterammina"\tab"disconorbis"\tab"discorbia"\cr
 \tab"discorbinella"\tab"discorbis"\tab"discorbitina"\cr
 \tab"edentostomina"\tab"elphidium"\tab"eoeponidella"\cr
 \tab"eponides"\tab"fissurina"\tab"fursenkoina"\cr
 \tab"glabratella"\tab"globocassidulina"\tab"hanzawaia"\cr
 \tab"hauerina"\tab"heronallenia"\tab"laevipeneroplis"\cr
 \tab"lepidodeuterammina"\tab"loxostomina"\tab"miliolinella"\cr
 \tab"mullinoides"\tab"mychostomina"\tab"neoconorbina"\cr
 \tab"nonionella"\tab"nonioninoides"\tab"pararotalia"\cr
 \tab"paratrochammina"\tab"patellina"\tab"planispirillina"\cr
 \tab"planorbulina"\tab"planulina"\tab"pyrgo"\cr
 \tab"quinqueloculina"\tab"rectocibicides"\tab"remaneicella"\cr
 \tab"reophax"\tab"rosalina"\tab"rotaliammina"\cr
 \tab"rotorbinella"\tab"rotorbis"\tab"sahulia"\cr
 \tab"sigmavirgulina"\tab"siphonina"\tab"sorites"\cr
 \tab"spirillina"\tab"spiroloculina"\tab"spirorbina"\cr
 \tab"tetrataxiella"\tab"textularia"\tab"triloculina"\cr
 \tab"triloculinella"\tab"trochammina"\tab"webbinella"\cr
 \tab"wiesnerella"
}
}
\keyword{`data.frame`}
