% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth}
\alias{depth}
\title{Import image object depth (Z)}
\usage{
depth(x, ...)
}
\arguments{
\item{xml_file}{image metadata file, \code{'.xml', '.tif_meta.xml'}}
}
\value{
A dataframe with the folowing information:
\itemize{
  \item {file} : {filename}
  \item {z_depth} : {measured focus range depth (z)}}
}
\description{
The function retrieve the object depth information from image metadata as source.
}
