% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{discorbinella}
\alias{discorbinella}
\title{Discorbinella size data}
\format{
A data frame with 318 rows and 7 variables:
\describe{
  \item{species}{species of genus Discorbinella from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{h}{test height in μm (15--72)}
  \item{d_one}{test minor diameter in  μm (67.69--197.43)}
  \item{radius}{test radius in  μm (33.85--98.71)}
  \item{area}{test surface area in μm2 (3599--30614)}
  \item{pco}{percent cell occupancy of the test (9.673--100)}
}
}
\usage{
discorbinella
}
\description{
A dataset containing the size and other attributes of species from genus Discorbinella.
 The variables are as follows:
}
\keyword{datasets}
