% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{exampleImage}
\alias{exampleImage}
\title{Foraminifeal photomicrographs for \code{forImage} examples}
\format{Image objects with variable dimension and resolution.}
\description{
This dataset contains image files of several foraminiferal genera.
These files are originally from previous work - Freitas et al. (2019)
}
\keyword{dataset}
