% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{laevipeneroplis}
\alias{laevipeneroplis}
\title{Laevipeneroplis size data}
\format{A data frame with 79 rows and 7 variables:
\describe{
  \item{species}{species of genus Laevipeneroplis from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{h}{test height in μm (40--633.9)}
  \item{d_one}{test minor diameter in  μm (47--624.2)}
  \item{area}{test surface area in μm2 (8827--306012)}
  \item{pop}{test protoplasm occupancy percentage (15.24--100)}
  \item{d_two}{test major diameter in  μm (169.2--544.0) - The NA's related to one species of this genus that don't required this measure}
}}
\usage{
laevipeneroplis
}
\description{
A dataset containing the size and other attributes of species from genus Laevipeneroplis.
 The variables are as follows:
}
\keyword{datasets}
