% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{loxostomina}
\alias{loxostomina}
\title{\emph{Loxostomina} size data}
\format{
A data frame with 31 rows and 7 variables:
\describe{
  \item{ind}{number of individuals by species}
  \item{d_one}{test minor diameter in μm (101.4--534.6)}
  \item{h}{test height in μm (28.0--80.0)}
  \item{d_two}{test major diameter in  μm (89.76--261.17)}
  \item{area}{test surface area in μm2 (6636--76089)}
  \item{pco}{percent cell occupancy of the test (9.88--100)}
}
}
\usage{
loxostomina
}
\description{
A dataset containing the size and other attributes of species from genus \emph{Loxostomina}.
 The variables are as follows:
}
\keyword{datasets}
