% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Foraminifera image measurement}
\usage{
measure(file, scale = NULL, ref_scale = NULL, pco = FALSE)
}
\arguments{
\item{file}{image file with or without metadata.}

\item{scale}{(optional) image physical pixel size (metric / pixel).}

\item{ref_scale}{(optional) reference scale available on image (in micrometers or millimeters). This scale should be inserted under the main object.}

\item{pco}{(optional) will assess proportion of cell occupancy inside the shell. Outlined proportion. This argument is s  till being tested and should be used with caution.}
}
\value{
An dataframe consisting of:
}
\description{
This function measures dimensions in photomicrographs.
See details \sQuote{Details}:
}
\examples{
2
}
\seealso{
\code{\link{bio.volume}}, \code{\link{volume.total}}
}
\author{
Thaise R. Freitas \email{thaisericardo.freitas@gmail.com}
}
