% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{patellina}
\alias{patellina}
\title{Patellina size data}
\format{A data frame with 79 rows and 6 variables:
\describe{
  \item{ind}{number of individuals by species}
  \item{h}{test height in μm (24--66)}
  \item{d_one}{test minor diameter in μm (84.84--247.46)}
  \item{radius}{test radius in  μm (42.42--123.73)}
  \item{area}{test surface area in μm2 (5653--87838)}
  \item{pco}{percent cell occupancy of the test (9.94--100)}
}}
\usage{
patellina
}
\description{
A dataset containing the size and other attributes of species from genus Patellina.
 The variables are as follows:
}
\keyword{datasets}
