% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{textularia}
\alias{textularia}
\title{Textularia size data}
\format{
A data frame with 84 rows and 7 variables:
\describe{
  \item{species}{species of genus Textularia from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{h}{test height in μm (113.6--1034.4)}
  \item{d_one}{test minor diameter in μm (40--306)}
  \item{d_two}{test major diameter in  μm (103.8--552.6)}
  \item{area}{test surface area in μm2 (8235--336929)}
  \item{pco}{percent cell occupancy of the test (10.46--100)}
}
}
\usage{
textularia
}
\description{
A dataset containing the size and other attributes of species from genus Textularia
 The variables are as follows:
}
\keyword{datasets}
