% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.total.R
\name{volume.total}
\alias{volume.total}
\title{Volumetric calculus of organisms}
\usage{
\method{volume}{total}(data, model, ...)
}
\arguments{
\item{data}{data frame containing size data. Size data parameters by model see.}

\item{model}{character informing geometric model to calculate volume, the models options are listed below:
\itemize{
  \item \code{'1hl'} : sphere
  \item \code{'2sl'} : half-sphere
  \item \code{'3hl'} : prolate spheroid
  \item \code{'4hl'} : cone or double cone
  \item \code{'6fs'} : paraboloid
  \item \code{'7fs'} : dome
  \item \code{'8hl'} : cylinder
  \item \code{'10hl'} : ellipsoid
  \item \code{'11fs'} : elliptic cone
  \item \code{'12v'} : cone + half ellipsoid
  \item \code{'13hlsl'} : gomphonemoid
  \item \code{'14hl'} : prism on elliptic base/elliptic cylinder
  \item \code{'15hl'} : half elliptic prism
  \item \code{'17fs'} : triangular dypyramid
  \item \code{'ahx'} : area x heigth

}}

\item{...}{other arguments.}
}
\value{
A dataframe or numeric object, consisting of calculated individual volume along with biovolume if the \code{pco} is informed.
}
\description{
The function calculates organisms volume based on geometric approximation.
}
\details{
These geometric models applied in this function are based in the selection made by Freitas et al. Which were mainly based and adapted from microalgae models developed by Hillebrand et al. (1999) - \code{('.hl')}, Sun and Liu (2003) - \code{('.sl')} and Vadrucci, Cabrini and Basset (2007) - \code{('.v')}. Some other formulas were adapted \code{('.fs')} to better calculate the test volume of benthic foraminifera.
 For more information, see \href{Freitas et al. (2019)}{Freitas et al. (2019)}
}
\examples{
#Ammonia size data
data("ammonia")

#calculate test volume
volume.total(ammonia, model = "10hl")


#with model argument in column (for more than 1 model)
am <- dplyr::mutate(ammonia, model = "10hl")
volume.total(am)

}
\references{
\itemize{
  \item Hillebrand, H., Dürselen, C.D., Kirschtel, D., Pollingher, U., & Zohary, T. (1999). Biovolume calculation for pelagic and benthic microalgae. \emph{Journal of Phycology}, 35(2), 403–424. \emph{doi:10.1046/j.1529-8817.1999.3520403.x}
  \item Sun, J., & Liu, D. (2003). Geometric models for calculating cell biovolume and surface area for phytoplankton. \emph{Journal of Plankton Research}, 25(11), 1331–1346. \emph{doi:10.1093/plankt/fbg096}
  \item Vadrucci, M. R., Cabrini, M., & Basset, A. (2007). Biovolume determination of phytoplankton guilds in transitional water ecosystems of Mediterranean Ecoregion. \emph{Transitional Waters Bulletin}, 2, 83–102. \emph{doi:10.1285/i1825229Xv1n2p83}
}
}
\seealso{
\code{\link{measure}}

\code{\link{bio.volume}}
}
\author{
Thaise R. Freitas \email{thaisericardo.freitas@gmail.com}
}
