% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{angulogerina}
\alias{angulogerina}
\title{Angulogerina size data}
\format{
A `data.frame` with 100 rows and 6 variables:
\describe{
  \item{ind}{number of individuals}
  \item{h}{test height in um (81.75--380.91)}
  \item{d_one}{test minor diameter in  um (34--156)}
  \item{d_two}{test major diameter in  um (66.04--237.34)}
  \item{area}{test surface area in um2 (4654--41915)}
  \item{pco}{percent cell occupancy of the test (6.818--100)}
}
}
\usage{
angulogerina
}
\description{
A `data.frame` containing the size and other attributes of species from genus \emph{Angulogerina}.
 The variables are as follows:
}
\keyword{datasets}
