% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{bolivina}
\alias{bolivina}
\title{Bolivina size data}
\format{
A `data.frame` with 628 rows and 7 variables:
\describe{
  \item{species}{species of genus Bolivina from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{d_one}{test minor diameter in um (72.37--502.12)}
  \item{h}{test height (thickness) in um (12.5--83.0)}
  \item{d_two}{test major diameter in  um (60.60--226.21)}
  \item{area}{test surface area in um2 (3187--72244)}
  \item{pco}{percent cell occupancy of the test (9.72--100)}
}
}
\usage{
bolivina
}
\description{
A `data.frame` containing the size and other attributes of species from genus \emph{Bolivina}.
 The variables are as follows:
}
\keyword{datasets}
