% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{cibicidoides}
\alias{cibicidoides}
\title{Cibicidoides size data}
\format{
A `data.frame` with 118 rows and 7 variables:
\describe{
  \item{species}{species of genus Cibicidoides from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{h}{test height in um (24--117)}
  \item{d_one}{test minor diameter in  um (76.46--266.85)}
  \item{radius}{test radius in  um (38.23--133.43)}
  \item{area}{test surface area in um2 (4591--55928)}
  \item{pco}{percent cell occupancy of the test (11.87--100)}
}
}
\usage{
cibicidoides
}
\description{
A `data.frame` containing the size and other attributes of species from genus \emph{Cibicidoides}.
 The variables are as follows:
}
\keyword{datasets}
