% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth}
\alias{depth}
\alias{depth.xml}
\title{Import image object depth (Z)}
\usage{
depth(x, ...)

\method{depth}{xml}(x, ...)
}
\arguments{
\item{x}{file containing metadata, if xml type: \code{'.xml', '.tif_meta.xml'}}

\item{...}{other arguments.}
}
\value{
A dataframe with the folowing information:
\itemize{
  \item {file} : {filename}
  \item {z_depth} : {measured focus range depth (z)}}
}
\description{
The function retrieves the object depth information from image metadata as source. Insert image file if metadata is embedded, or separate metadata file if not.
}
