% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{exampleImage}
\alias{exampleImage}
\title{Foraminiferal photomicrograph for \code{forImage} examples}
\format{
Image objects with variable dimension and resolution.
}
\description{
Contains image file of one foraminifer.
These files are originally from previous work - Freitas and others (2019)
}
\keyword{`data.frame`}
