% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Foraminifera image measurement}
\usage{
measure(filePath, ...)
}
\arguments{
\item{filePath}{path containing the photomicrographs along with their metadata.}

\item{...}{other arguments.}
}
\value{
An dataframe consisting of:
}
\description{
The function identifies forams in photomicrographs and measure the individuals dimensions.
See details \sQuote{Details}:
}
\examples{
2
}
\seealso{
\code{\link{bio.volume}}, \code{\link{volume.total}}
}
\author{
Thaise R. Freitas \email{thaisericardo.freitas@gmail.com}
}
