% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{nonionella}
\alias{nonionella}
\title{Nonionella size data}
\format{
A `data.frame` with 208 rows and 7 variables:
\describe{
  \item{species}{species of genus Nonionella from which raw data were gathered}
  \item{ind}{number of individuals by species}
  \item{d_one}{test minor diameter in um (93.39--564.98)}
  \item{h}{test height in um (19.50--294.00)}
  \item{d_two}{test major diameter in  um (82.99--547.67)}
  \item{area}{test surface area in um2 (5783--222371)}
  \item{pco}{percent cell occupancy of the test (15.51--100)}
}
}
\usage{
nonionella
}
\description{
A `data.frame` containing the size and other attributes of species from genus \emph{Nonionella}.
 The variables are as follows:
}
\keyword{datasets}
