% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{rectocibicides}
\alias{rectocibicides}
\title{Rectocibicides size data}
\format{
A `data.frame` with 199 rows and 4 variables:
\describe{
  \item{ind}{number of individuals}
  \item{h}{test height in um (22.00--101.50)}
  \item{area}{test surface area in um2 (3450--674280)}
  \item{pco}{percent cell occupancy of the test (10--100)}

}
}
\usage{
rectocibicides
}
\description{
A `data.frame` containing the size and other attributes of genus \emph{Rectocibicides}.
 The variables are as follows:
}
\keyword{datasets}
