% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foram.R
\docType{data}
\name{spirillina}
\alias{spirillina}
\title{Spirillina size data}
\format{
A `data.frame` with 42 rows and 6 variables:
\describe{
  \item{ind}{number of individuals by species}
  \item{h}{test height in um (12--44)}
  \item{d_one}{test minor diameter in um (73.71--105.14)}
  \item{radius}{test radius in  um (36.86--123.73)}
  \item{area}{test surface area in um2 (4268--34730)}
  \item{pco}{percent cell occupancy of the test (7.10--100)}
}
}
\usage{
spirillina
}
\description{
A `data.frame` containing the size and other attributes of species from genus \emph{Spirillina}.
 The variables are as follows:
}
\keyword{datasets}
